refactor: Phase 5 — extract SDNA and SHACL generation from decorators.ts
- Create sdna.ts (~170 lines) with buildSDNA() for Prolog SDNA generation
- Create shacl-gen.ts (~292 lines) with buildSHACL() for SHACL shape generation
- Move makeRandomId() to util.ts, re-export from decorators.ts for compat
- Model decorator now delegates to buildSDNA/buildSHACL (no circular deps)
- decorators.ts reduced from 1484 to 1066 lines
- Clean up unused imports (SHACLShape, compileWhereClause, etc)
- All 119 tests pass (115 passed, 4 skipped)
