+++ core/src/model/decorators.ts
@@ -1,8 +1,9 @@
-import { capitalize, propertyNameToSetterName, singularToPlural, stringifyObjectLiteral } from "./util";
-import { SHACLShape, SHACLPropertyShape, ConformanceCondition } from "../shacl/SHACLShape";
+import { capitalize, propertyNameToSetterName } from "./util";
+import { ConformanceCondition } from "../shacl/SHACLShape";
 import { escapeSurrealString } from "../utils";
 import type { Where } from "./types";
-import { compileWhereClause } from "./surreal-utils";
+import { buildSDNA } from "./sdna";
+import { buildSHACL } from "./shacl-gen";
 
 // ============================================================================
 // WeakMap-based metadata registry
@@ -209,18 +210,10 @@ export function instanceToSerializable(instance: any): Record<string, any> {
 }
 
 /**
- * Generate a random identifier string (lowercase alpha).
  * Generate a random identifier string of the given length (lowercase alpha).
+ * @deprecated Import from `./util` directly. Re-exported here for backward compatibility.
  */
-export function makeRandomId(length: number): string {
-    let result = '';
-    const characters = 'abcdefghijklmnopqrstuvwxyz';
-    const charactersLength = characters.length;
-    for (let i = 0; i < length; i++) {
-        result += characters.charAt(Math.floor(Math.random() * charactersLength));
-    }
-    return result;
-}
+export { makeRandomId } from "./util";
 
 export class PerspectiveAction {
     action: string
@@ -531,432 +524,22 @@ export function Model(opts: ModelConfig) {
         target.className = opts.name;
 
         target.generateSDNA = function() {
-            let sdna = ""
-            let subjectName = opts.name
-            let obj = target.prototype;
-
-            let uuid = makeRandomId(8)
-
-            sdna += `subject_class("${subjectName}", ${uuid}).\n`
-
-
-            let classRemoverActions = []
-
-            let constructorActions = []
-            if(obj.subjectConstructor && obj.subjectConstructor.length) {
-                constructorActions = constructorActions.concat(obj.subjectConstructor)
-            }
-
-            let instanceConditions = []
-            if(obj.isSubjectInstance && obj.isSubjectInstance.length) {
-                instanceConditions = instanceConditions.concat(obj.isSubjectInstance)
-            }
-
-            let propertiesCode = []
-            let properties = getPropertiesMetadata(target)
-            for(let property in properties) {
-                let propertyCode = `property(${uuid}, "${property}").\n`
-
-                let { through, initial, required, resolveLanguage, writable, flag, prologGetter, prologSetter, local } = properties[property]
-
-                if(resolveLanguage) {
-                    propertyCode += `property_resolve(${uuid}, "${property}").\n`
-                    propertyCode += `property_resolve_language(${uuid}, "${property}", "${resolveLanguage}").\n`
-                }
-
-                if(prologGetter) {
-                    propertyCode += `property_getter(${uuid}, Base, "${property}", Value) :- ${prologGetter}.\n`
-                } else if(through) {
-                    propertyCode += `property_getter(${uuid}, Base, "${property}", Value) :- triple(Base, "${through}", Value).\n`
-
-                    if(required) {
-                        if(flag) {
-                            instanceConditions.push(`triple(Base, "${through}", "${initial}")`)
-                        } else {
-                            instanceConditions.push(`triple(Base, "${through}", _)`)
-                        }
-                    }
-                }
-
-                if(prologSetter) {
-                    propertyCode += `property_setter(${uuid}, "${property}", Actions) :- ${prologSetter}.\n`
-                } else if (writable && through) {
-                    let setter = obj[propertyNameToSetterName(property)]
-                    if(typeof setter === "function") {
-                        let action = [{
-                            action: "setSingleTarget",
-                            source: "this",
-                            predicate: through,
-                            target: "value",
-                            ...(local && { local: true })
-                        }]
-                        propertyCode += `property_setter(${uuid}, "${property}", '${stringifyObjectLiteral(action)}').\n`
-                    }
-                }
-
-                propertiesCode.push(propertyCode)
-
-                // Auto-derive effectiveInitial for required, writable, non-flag properties
-                // so constructor/destructor always handle the property link.
-                // Optional properties (required: false) are excluded — they should
-                // remain unset (undefined) until explicitly assigned.
-                const effectiveInitial = initial
-                    ?? (required && writable && !flag && through ? "literal://string:" : undefined);
-
-                if(effectiveInitial) {
-                    constructorActions.push({
-                        action: "addLink",
-                        source: "this",
-                        predicate: through,
-                        target: effectiveInitial,
-                    })
-
-                    classRemoverActions.push({
-                        action: "removeLink",
-                        source: "this",
-                        predicate: through,
-                        target: "*",
-                    })
-                }
-            }
-
-            let relationsCode = []
-            const allRelationsMeta = getRelationsMetadata(target)
-            const relations = Object.fromEntries(
-                Object.entries(allRelationsMeta).filter(([, r]) => r.kind === 'hasMany' || r.kind === 'belongsToMany')
-            )
-            for(let relation in relations) {
-                let relationCode = `collection(${uuid}, "${relation}").\n`
-
-                let { predicate: through, local} = relations[relation]
-
-                if(through) {
-                    relationCode += `collection_getter(${uuid}, Base, "${relation}", List) :- findall(C, triple(Base, "${through}", C), List).\n`
-
-                    let relationAdderAction = [{
-                        action: "addLink",
-                        source: "this",
-                        predicate: through,
-                        target: "value",
-                        ...(local && { local: true })
-                    }]
-
-                    let relationRemoverAction = [{
-                        action: "removeLink",
-                        source: "this",
-                        predicate: through,
-                        target: "value",
-                    }]
-
-                    let relationSetterAction = [{
-                        action: "collectionSetter",
-                        source: "this",
-                        predicate: through,
-                        target: "value",
-                        ...(local && { local: true })
-                    }]
-                    relationCode += `collection_adder(${uuid}, "${relation}", '${stringifyObjectLiteral(relationAdderAction)}').\n`
-                    relationCode += `collection_remover(${uuid}, "${relation}", '${stringifyObjectLiteral(relationRemoverAction)}').\n`
-                    relationCode += `collection_setter(${uuid}, "${relation}", '${stringifyObjectLiteral(relationSetterAction)}').\n`
-                }
-
-                relationsCode.push(relationCode)
-            }
-
-            let subjectContructorJSONString = stringifyObjectLiteral(constructorActions)
-            sdna += `constructor(${uuid}, '${subjectContructorJSONString}').\n`
-            if(instanceConditions.length > 0) {
-                let instanceConditionProlog = instanceConditions.join(", ")
-                sdna += `instance(${uuid}, Base) :- ${instanceConditionProlog}.\n`
-                sdna += "\n"
-            }
-            sdna += `destructor(${uuid}, '${stringifyObjectLiteral(classRemoverActions)}').\n`
-            sdna += "\n"
-            sdna += propertiesCode.join("\n")
-            sdna += "\n"
-            sdna += relationsCode.join("\n")
-
-            return {
-                sdna,
-                name: subjectName
-            }
+            return buildSDNA(
+                opts.name,
+                target.prototype,
+                getPropertiesMetadata(target),
+                getRelationsMetadata(target),
+            );
         }
 
-        // Generate SHACL shape (W3C standard + AD4M action definitions)
         target.generateSHACL = function() {
-            const subjectName = opts.name;
-            const obj = target.prototype;
-
-            // Determine namespace from first property or relation, or use default
-            let namespace = "ad4m://";
-            const properties = getPropertiesMetadata(target);
-            const allRelationsMeta2 = getRelationsMetadata(target);
-            const relations = Object.fromEntries(
-                Object.entries(allRelationsMeta2).filter(([, r]) => r.kind === 'hasMany' || r.kind === 'belongsToMany')
+            return buildSHACL(
+                opts.name,
+                target,
+                getPropertiesMetadata(target),
+                getRelationsMetadata(target),
+                buildConformanceFilter,
             );
-            
-            // Try properties first
-            if (Object.keys(properties).length > 0) {
-                const firstProp = properties[Object.keys(properties)[0]];
-                if (firstProp.through) {
-                    // Extract namespace from through predicate (e.g., "recipe://name" -> "recipe://")
-                    const match = firstProp.through.match(/^([^:]+:\/\/)/);
-                    if (match) {
-                        namespace = match[1];
-                    }
-                }
-            } 
-            // Fall back to relations if no properties
-            else if (Object.keys(relations).length > 0) {
-                const firstRel = relations[Object.keys(relations)[0]];
-                if (firstRel.predicate) {
-                    const match = firstRel.predicate.match(/^([^:]+:\/\/)/);
-                    if (match) {
-                        namespace = match[1];
-                    }
-                }
-            }
-
-            // Create SHACL shape
-            const shapeUri = `${namespace}${subjectName}Shape`;
-            const targetClass = `${namespace}${subjectName}`;
-            const shape = new SHACLShape(shapeUri, targetClass);
-
-            // Detect @Model inheritance — if the parent class also has
-            // generateSHACL it is itself a @Model and we reference its shape
-            // via sh:node so SHACL validators can walk the hierarchy.
-            const parentCtor = Object.getPrototypeOf(target);
-            if (parentCtor && typeof parentCtor.generateSHACL === 'function') {
-                const parentSHACL = parentCtor.generateSHACL();
-                if (parentSHACL?.shape?.nodeShapeUri) {
-                    shape.addParentShape(parentSHACL.shape.nodeShapeUri);
-                }
-            }
-
-            // === Extract Constructor Actions (same logic as generateSDNA) ===
-            let constructorActions = [];
-            if(obj.subjectConstructor && obj.subjectConstructor.length) {
-                constructorActions = constructorActions.concat(obj.subjectConstructor);
-            }
-
-            // === Extract Destructor Actions ===
-            let destructorActions = [];
-            
-            // Convert properties to SHACL property shapes
-            for (const propName in properties) {
-                const propMeta = properties[propName];
-
-                if (!propMeta.through) continue; // Skip properties without predicates
-
-                const propShape: SHACLPropertyShape = {
-                    name: propName, // Property name for generating named URIs
-                    path: propMeta.through,
-                };
-
-                // Determine datatype from initial value or resolveLanguage
-                if (propMeta.resolveLanguage === "literal") {
-                    // If it resolves via literal language, it's likely a string
-                    propShape.datatype = "xsd://string";
-                } else if (propMeta.initial) {
-                    // Try to infer from initial value type
-                    const initialType = typeof obj[propName];
-                    if (initialType === "number") {
-                        propShape.datatype = "xsd://integer";
-                    } else if (initialType === "boolean") {
-                        propShape.datatype = "xsd://boolean";
-                    } else if (initialType === "string") {
-                        propShape.datatype = "xsd://string";
-                    }
-                }
-
-                // Cardinality constraints
-                if (propMeta.required) {
-                    propShape.minCount = 1;
-                }
-
-                // Single-valued properties get maxCount 1
-                // (relations are handled separately below)
-                propShape.maxCount = 1;
-
-                // Flag properties have fixed value
-                if (propMeta.flag && propMeta.initial) {
-                    propShape.hasValue = propMeta.initial;
-                }
-
-                // AD4M-specific metadata
-                if (propMeta.local !== undefined) {
-                    propShape.local = propMeta.local;
-                }
-
-                if (propMeta.writable !== undefined) {
-                    propShape.writable = propMeta.writable;
-                }
-
-                if (propMeta.resolveLanguage) {
-                    propShape.resolveLanguage = propMeta.resolveLanguage;
-                }
-
-                // === Extract Setter Actions (same logic as generateSDNA) ===
-                if (propMeta.prologSetter) {
-                    // Custom Prolog setter defined - not yet supported in SHACL
-                    console.warn(
-                        `[SHACL Generation] Custom Prolog setter for property '${propName}' in class '${subjectName}' is not yet supported. ` +
-                        `The property will be created without setter actions. Consider using standard writable properties or provide explicit SHACL JSON.`
-                    );
-                    // TODO: Parse custom Prolog setter to extract actions
-                } else if (propMeta.writable && propMeta.through) {
-                    let setter = obj[propertyNameToSetterName(propName)];
-                    if (typeof setter === "function") {
-                        propShape.setter = [{
-                            action: "setSingleTarget",
-                            source: "this",
-                            predicate: propMeta.through,
-                            target: "value",
-                            ...(propMeta.local && { local: true })
-                        }];
-                    }
-                }
-
-                // Add to constructor actions if property has initial value
-                // Auto-derive effectiveInitial for required, writable, non-flag properties.
-                // Optional properties (required: false) are excluded — they should
-                // remain unset (undefined) until explicitly assigned.
-                const effectiveInitial = propMeta.initial
-                    ?? (propMeta.required && propMeta.writable && !propMeta.flag && propMeta.through
-                        ? "literal://string:" : undefined);
-
-                if (effectiveInitial) {
-                    constructorActions.push({
-                        action: "addLink",
-                        source: "this",
-                        predicate: propMeta.through,
-                        target: effectiveInitial,
-                    });
-
-                    // Add to destructor actions
-                    destructorActions.push({
-                        action: "removeLink",
-                        source: "this",
-                        predicate: propMeta.through,
-                        target: "*",
-                    });
-                }
-
-                shape.addProperty(propShape);
-            }
-            
-            // Convert relations to SHACL property shapes
-            // (relations variable already declared above for namespace inference)
-            for (const relName in relations) {
-                const relMeta = relations[relName];
-                
-                if (!relMeta.predicate) continue;
-                
-                const relShape: SHACLPropertyShape = {
-                    name: relName, // Relation name for generating named URIs
-                    path: relMeta.predicate,
-                    // Relations have no maxCount (unlimited)
-                    // minCount defaults to 0 (optional)
-                };
-                
-                // Determine node kind — relations typically contain IRIs
-                relShape.nodeKind = 'IRI';
-                
-                // AD4M-specific metadata
-                if (relMeta.local !== undefined) {
-                    relShape.local = relMeta.local;
-                }
-
-                // === Extract Relation Actions (adder/remover) ===
-                // Adder action - adds a link to the relation
-                relShape.adder = [{
-                    action: "addLink",
-                    source: "this",
-                    predicate: relMeta.predicate,
-                    target: "value",
-                    ...(relMeta.local && { local: true })
-                }];
-
-                // Remover action - removes a link from the relation
-                relShape.remover = [{
-                    action: "removeLink",
-                    source: "this",
-                    predicate: relMeta.predicate,
-                    target: "value",
-                    ...(relMeta.local && { local: true })
-                }];
-
-                // === Build Getter (conformance filter) ===
-                // Priority chain:
-                // 1. Explicit getter string → use verbatim
-                // 2. `where` clause → compile DSL to SurrealQL getter
-                // 3. target + filter !== false → auto-derive from shape
-                if (relMeta.getter) {
-                    relShape.getter = relMeta.getter;
-                } else if (relMeta.where) {
-                    // Compile Where DSL → SurrealQL getter
-                    try {
-                        const TargetClass = relMeta.target?.();
-                        const targetMetadata = TargetClass
-                            ? (TargetClass as any).getModelMetadata?.() ?? null
-                            : null;
-
-                        const conditions = compileWhereClause(
-                            relMeta.where,
-                            targetMetadata,
-                        );
-
-                        if (conditions.length > 0) {
-                            const escapedPredicate = escapeSurrealString(relMeta.predicate);
-                            relShape.getter = `(->link[WHERE predicate = '${escapedPredicate}'].out[WHERE ${conditions.join(' AND ')}].uri)`;
-                        }
-                    } catch (e) {
-                        // Target metadata may not be available yet
-                    }
-                } else if (relMeta.target && relMeta.filter !== false) {
-                    try {
-                        const TargetClass = relMeta.target();
-                        const filter = buildConformanceFilter(relMeta.predicate, TargetClass);
-                        if (filter) {
-                            relShape.getter = filter.getter;
-                            relShape.conformanceConditions = filter.conformanceConditions;
-                        }
-                    } catch (e) {
-                        // Target class may not be available yet — getter will be
-                        // absent and runtime will fall back to unfiltered get_links
-                    }
-                }
-
-                // === sh:class — target shape reference ===
-                // Always set when a target model is declared, regardless of filter.
-                // This enables typed construction on the Rust/MCP side.
-                if (relMeta.target) {
-                    try {
-                        const TargetClass = relMeta.target();
-                        const targetSHACL = (TargetClass as any).generateSHACL?.();
-                        if (targetSHACL?.shape?.nodeShapeUri) {
-                            relShape.class = targetSHACL.shape.nodeShapeUri;
-                        }
-                    } catch (e) {
-                        // Target class may not be available yet
-                    }
-                }
-
-                shape.addProperty(relShape);
-            }
-
-            // Always set constructor and destructor actions on the shape, even
-            // when empty.  An empty array serialises to `literal://string:[]`
-            // which the Rust executor parses as a valid (no-op) command list,
-            // avoiding "No SHACL constructor found" errors for models whose
-            // properties are all optional and have no @Flag.
-            shape.setConstructorActions(constructorActions);
-            shape.setDestructorActions(destructorActions);
-
-            return {
-                shape,
-                name: subjectName
-            };
         }
 
         Object.defineProperty(target, 'type', {configurable: true});
